move Document and Annotation types back to src.document.types
