+++ configs/dataset/_add_candidate_relations.yaml
@@ -3,6 +3,6 @@ add_candidate_relations:
   function: src.utils.document.add_candidate_relations
   # the result_type is required because add_reversed_relations is a generic document conversion function
   # that does not define a specific return values
-  result_type: src.types.document.DocumentWithEntitiesRelationsAndLabeledPartitions
+  result_type: src.document.types.DocumentWithEntitiesRelationsAndLabeledPartitions
   #partition_layer: partitions
   #max_distance: 177
