+++ configs/dataset/_remove_overlapping_entities.yaml
@@ -1,4 +1,4 @@
 remove_overlapping_entities:
   _processor_: src.utils.dataset.convert_documents
   function: src.utils.document.remove_overlapping_entities
-  result_type: src.types.document.DocumentWithEntitiesRelationsAndLabeledPartitions
+  result_type: src.document.types.DocumentWithEntitiesRelationsAndLabeledPartitions
